project/requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
isodate==0.6.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2021.1
rdflib==5.0.0
six==1.16.0
SPARQLWrapper==1.8.5
sqlparse==0.4.1
